+++ libs/spark/src/Unstable_TextField/Unstable_TextField.tsx
@@ -16,7 +16,6 @@ import Unstable_FormHelperText, {
   Unstable_FormHelperTextProps,
 } from '../Unstable_FormHelperText';
 import Unstable_Select, { Unstable_SelectProps } from '../Unstable_Select';
-import { useId } from '../utils';
 
 export interface Unstable_TextFieldProps
   extends Omit<
@@ -135,10 +134,6 @@ export interface Unstable_TextFieldProps
    */
   SelectProps?: Partial<Unstable_SelectProps>;
   /**
-   * If `true`, the input will indicate a success.
-   */
-  success?: boolean;
-  /**
    * The content of the `endAdornment` (an `InputAdornment`), usually an `Icon`, `IconButton`, or `string`.
    */
   trailingEl?: ReactNode;
@@ -164,7 +159,6 @@ const Unstable_TextField = forwardRef<unknown, Unstable_TextFieldProps>(
       FormHelperTextProps,
       fullWidth = false,
       helperText,
-      id: idProp,
       FormLabelProps,
       inputProps,
       InputProps,
@@ -182,15 +176,13 @@ const Unstable_TextField = forwardRef<unknown, Unstable_TextFieldProps>(
       minRows,
       select = false,
       SelectProps,
-      size = 'medium',
+      size,
       success,
       trailingEl,
       type,
       value,
       ...other
     } = props;
-
-    const id = useId(idProp);
 
     if (process.env.NODE_ENV !== 'production') {
       if (select && !children) {
@@ -210,11 +202,8 @@ const Unstable_TextField = forwardRef<unknown, Unstable_TextFieldProps>(
       InputMore['aria-describedby'] = undefined;
     }
 
-    const helperTextId = helperText && id ? `${id}-helper-text` : undefined;
-    const labelId = label && id ? `${id}-label` : undefined;
     const InputElement = (
       <Unstable_Input
-        aria-describedby={helperTextId}
         autoComplete={autoComplete}
         autoFocus={autoFocus}
         disabled={disabled}
@@ -225,12 +214,9 @@ const Unstable_TextField = forwardRef<unknown, Unstable_TextFieldProps>(
         name={name}
         maxRows={maxRows}
         minRows={minRows}
-        size={size}
-        success={success}
         trailingEl={trailingEl}
         type={type}
         value={value}
-        id={id}
         inputRef={inputRef}
         onBlur={onBlur}
         onChange={onChange}
@@ -250,29 +236,15 @@ const Unstable_TextField = forwardRef<unknown, Unstable_TextFieldProps>(
         ref={(ref as unknown) as RefObject<HTMLDivElement>}
         required={required}
         size={size}
+        success={success}
         {...other}
       >
         {label ? (
-          <Unstable_FormLabel
-            htmlFor={id}
-            id={labelId}
-            size={size}
-            {...FormLabelProps}
-          >
-            {label}
-          </Unstable_FormLabel>
+          <Unstable_FormLabel {...FormLabelProps}>{label}</Unstable_FormLabel>
         ) : null}
 
         {select ? (
-          <Unstable_Select
-            aria-describedby={helperTextId}
-            id={id}
-            labelId={labelId}
-            size={size}
-            value={value}
-            input={InputElement}
-            {...SelectProps}
-          >
+          <Unstable_Select value={value} input={InputElement} {...SelectProps}>
             {children}
           </Unstable_Select>
         ) : (
@@ -280,11 +252,7 @@ const Unstable_TextField = forwardRef<unknown, Unstable_TextFieldProps>(
         )}
 
         {helperText ? (
-          <Unstable_FormHelperText
-            id={helperTextId}
-            size={size}
-            {...FormHelperTextProps}
-          >
+          <Unstable_FormHelperText {...FormHelperTextProps}>
             {helperText}
           </Unstable_FormHelperText>
         ) : null}